Extract generator to outside of local C.CSS
