+++ projects/plugins/boost/app/data-sync/Critical_CSS_Meta_Entry.php
@@ -2,7 +2,7 @@
 namespace Automattic\Jetpack_Boost\Data_Sync;
 
 use Automattic\Jetpack\WP_JS_Data_Sync\Contracts\Entry_Can_Get;
-use Automattic\Jetpack_Boost\Modules\Optimizations\Critical_CSS\Generator;
+use Automattic\Jetpack_Boost\Lib\Critical_CSS\Generator;
 
 class Critical_CSS_Meta_Entry implements Entry_Can_Get {
 	public function get() {
